Clean up tests to refactor out fixtures
